cosplay-ai-v1/requirements.txt
# FastAPI and web server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# Google Cloud and AI Platform
google-cloud-aiplatform>=1.30.0
google-cloud-storage>=2.8.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.6.0
numpy>=1.21.0

# Utilities
python-dotenv>=0.19.0
pydantic>=2.0.0
httpx>=0.24.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
flake8>=5.0.0

# Optional: Database (for future use)
# sqlalchemy>=2.0.0
# alembic>=1.8.0